Add GitHub Actions test workflow.
