+++ .coveragerc
@@ -1,6 +1,3 @@
 [run]
 branch = True
-include =
-    widget_tweaks/*
-omit =
-    widget_tweaks/tests.py
+source = widget_tweaks
