--- .travis.yml
@@ -1,39 +1,0 @@
-dist: xenial
-sudo: false
-language: python
-cache: pip
-python:
-  - 3.6
-  - 3.7
-  - 3.8
-env:
-  - DJANGO=1.11
-  - DJANGO=2.2
-  - DJANGO=3.0
-  - DJANGO=master
-install: pip install tox-travis codecov
-script: tox
-after_success: codecov
-jobs:
-  fast_finish: true
-  allow_failures:
-    - env: DJANGO=master
-  include:
-    - python: 3.6
-      env: TOXENV=qa
-    - stage: deploy
-      env:
-      python: 3.6
-      script: skip
-      deploy:
-        provider: pypi
-        user: jazzband
-        server: https://jazzband.co/projects/django-widget-tweaks/upload
-        distributions: sdist bdist_wheel
-        password:
-          secure: hW8AcAymkogDbN2ZZ6KQBynTRMap30xAL33HtG4WDkh1+maGpkVAXn/HvLqrRkO/bh1h85uVS2QhRZifX3eKxxNc83Rp6YACQMobW2xBPrXdMIxl5t1t7R8e14pW6XWb6Xjy/zdmZd/b6m1Q3K1a6GvNqTxkfylVfAgJ8gBfRAs=
-        skip_existing: true
-        on:
-          tags: true
-          repo: jazzband/django-widget-tweaks
-          python: 3.6
