+++ tests/settings.py
@@ -1,0 +1,9 @@
+INSTALLED_APPS = ["widget_tweaks"]
+
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+    },
+]
+
+SECRET_KEY = "spam-eggs"
