+++ tests/tests.py
@@ -1,77 +1,6 @@
-import string
+from unittest import TestCase
 
-try:
-    from unittest import TestCase, skipIf
-except ImportError:
-    from unittest2 import TestCase, skipIf
-
-from django import VERSION
-from django import forms
-from django.forms import Form, CharField, TextInput
-from django.template import Template, Context
-
-try:
-    from django.forms import SelectDateWidget  # django >= 2.0
-except ImportError:
-    from django.forms.extras.widgets import SelectDateWidget  # django < 2.0
-
-
-# ==============================
-#       Testing helpers
-# ==============================
-
-
-class MyForm(Form):
-    """
-    Test form. If you want to test rendering of a field,
-    add it to this form and use one of 'render_...' functions
-    from this module.
-    """
-
-    simple = CharField()
-    with_attrs = CharField(widget=TextInput(attrs={"foo": "baz", "egg": "spam"}))
-    with_cls = CharField(widget=TextInput(attrs={"class": "class0"}))
-    date = forms.DateField(widget=SelectDateWidget(attrs={"egg": "spam"}))
-
-
-def render_form(text, form=None, **context_args):
-    """
-    Renders template ``text`` with widget_tweaks library loaded
-    and MyForm instance available in context as ``form``.
-    """
-    tpl = Template("{% load widget_tweaks %}" + text)
-    context_args.update({"form": MyForm() if form is None else form})
-    context = Context(context_args)
-    return tpl.render(context)
-
-
-def render_field(field, template_filter, params, *args, **kwargs):
-    """
-    Renders ``field`` of MyForm with filter ``template_filter`` applied.
-    ``params`` are filter arguments.
-
-    If you want to apply several filters (in a chain),
-    pass extra ``template_filter`` and ``params`` as positional arguments.
-
-    In order to use custom form, pass form instance as ``form``
-    keyword argument.
-    """
-    filters = [(template_filter, params)]
-    filters.extend(zip(args[::2], args[1::2]))
-    filter_strings = ['|%s:"%s"' % (f[0], f[1]) for f in filters]
-    render_field_str = "{{ form.%s%s }}" % (field, "".join(filter_strings))
-    return render_form(render_field_str, **kwargs)
-
-
-def render_field_from_tag(field, *attributes):
-    """
-    Renders MyForm's field ``field`` with attributes passed
-    as positional arguments.
-    """
-    attr_strings = [" %s" % f for f in attributes]
-    tpl = string.Template("{% render_field form.$field$attrs %}")
-    render_field_str = tpl.substitute(field=field, attrs="".join(attr_strings))
-    return render_form(render_field_str)
+from .forms import render_field, render_field_from_tag, render_form, MyForm
 
 
 def assertIn(value, obj):
@@ -182,16 +111,6 @@ class CustomizedWidgetTest(TestCase):
         assertNotIn('foo="baz"', res)
         assertIn('egg="spam"', res)
 
-    # XXX can be dropped once 1.8 is not supported
-    @skipIf(
-        VERSION < (1, 11, 0, "final", 0),
-        "see https://code.djangoproject.com/ticket/16754",
-    )
-    def test_selectdatewidget(self):
-        res = render_field("date", "attr", "foo:bar")
-        assertIn('egg="spam"', res)
-        assertIn('foo="bar"', res)
-
     def test_attr_chaining(self):
         res = render_field("with_attrs", "attr", "foo:bar", "attr", "bar:baz")
         assertIn('foo="bar"', res)
@@ -278,16 +197,6 @@ class RenderFieldTagCustomizedWidgetTest(TestCase):
         assertIn('foo="bar"', res)
         assertNotIn('foo="baz"', res)
         assertIn('egg="spam"', res)
-
-    # XXX can be dropped once 1.8 is not supported
-    @skipIf(
-        VERSION < (1, 11, 0, "final", 0),
-        "see https://code.djangoproject.com/ticket/16754",
-    )
-    def test_selectdatewidget(self):
-        res = render_field_from_tag("date", 'foo="bar"')
-        assertIn('egg="spam"', res)
-        assertIn('foo="bar"', res)
 
     def test_multiple_attrs(self):
         res = render_field_from_tag("with_attrs", 'foo="bar"', 'bar="baz"')
@@ -425,7 +334,7 @@ class RenderFieldTagFieldReuseTest(TestCase):
         self.assertEqual(res.count("c_1"), 1)
         self.assertEqual(res.count("c_2"), 1)
 
-    def test_field_double_rendering_simple(self):
+    def test_field_double_rendering_simple_again(self):
         res = render_form('{% render_field form.simple foo="bar" v-model="username" %}')
         self.assertEqual(res.count('v-model="username"'), 1)
