+++ tox.ini
@@ -1,37 +1,36 @@
 [tox]
 envlist =
-    py{36,37,38}-django111
-    py{36,37,38}-django22
-    py{36,37,38}-django30
-    py{36,37,38}-djangomaster
+    py{36,37,38}-dj{22,30,31,master}
+    qa
 
-[travis]
+[gh-actions]
 python =
-    3.6: py36
+    3.6: py36, qa
     3.7: py37
     3.8: py38
+    3.9: py39
 
-[travis:env]
+[gh-actions:env]
 DJANGO =
-    1.11: django111
-    2.2: django22
-    3.0: django30
-    master: djangomaster
+    2.2: dj22
+    3.0: dj30
+    3.1: dj31
+    master: djmaster
 
 [testenv]
 deps =
     coverage
-    django111: django>=1.11,<2.0
-    django22: django>=2.2,<2.3
-    django30: django>=3.0,<3.1
-    djangomaster: https://github.com/django/django/archive/master.tar.gz
+    dj22: django>=2.2,<2.3
+    dj30: django>=3.0,<3.1
+    dj31: django>=3.1,<3.2
+    djmaster: https://github.com/django/django/archive/master.tar.gz
 usedevelop = True
 ignore_outcome =
     djangomaster: True
 commands =
-    ./runtests.py {posargs}
-setenv =
-    PYTHONDONTWRITEBYTECODE=1
+    {envbindir}/coverage run {envbindir}/django-admin test -v2 --settings=tests.settings
+    {envbindir}/coverage report -m
+    {envbindir}/coverage xml
 
 [testenv:qa]
 deps =
