+++ widget_tweaks/__init__.py
@@ -1,0 +1,7 @@
+from pkg_resources import get_distribution, DistributionNotFound
+
+try:
+    __version__ = get_distribution("django-widget-tweaks").version
+except DistributionNotFound:
+    # package is not installed
+    __version__ = None
